Flat file import will soon be back to App Store to support Mendix 6
The author of 'Flat file importer' contacted me that 'Flat file importer' has been improved for Mendix 6 and is now in queue for App Store.  The fixes are different for Mendix 6 and 7, and their mpk files appear in App Store at different timings.
Thus, I would like to redeem it in the docs.  I have already checked it working in Mendix 6.
The docs should be updated as soon as a new 'Flat file importer' shows up in App Store.
Thanks.